// packages/api-client/src/generated-types/TopicPerformanceQueryParams.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs).
// Relative imports converted to absolute imports by generate-types recipe. DO NOT EDIT.
import type { SortOrder } from "@/generated-types/SortOrder";
import type { TopicPerformanceSortBy } from "@/generated-types/TopicPerformanceSortBy";

export type TopicPerformanceQueryParams = {
  subject_id?: string;
  chapter_id?: string;
  topic_id?: string;
  page?: number;
  limit?: number;
  search?: string;
  sort_by?: TopicPerformanceSortBy;
  sort_order?: SortOrder;
};